refactor: extract session snapshot folds
